+++ packages/orm/angel_orm_generator/CHANGELOG.md
@@ -1,8 +1,14 @@
 # Change Log
 
+## 4.1.1
+
+* Fixed `NumericSqlExpressionBuilder` to handle nullable field
+* Fixed `@belongsTo` code generation
+* Fixed `copyFrom` to handle nullable relationship
+
 ## 4.1.0
 
-* Upgraded to support major `analyzer` 2.0.0 release
+* Upgraded to support `analyzer` 2.0.0 major release
 
 ## 4.0.2